+++ src/AvroConvert/AvroObjectServices/BuildSchema/AvroSerializerSettings.cs
@@ -24,7 +24,7 @@ namespace SolTechnology.Avro.AvroObjectServices.BuildSchema
     /// <summary>
     ///     Specifies Avro serializer settings.
     /// </summary>
-    internal sealed class AvroSerializerSettings : IEquatable<AvroSerializerSettings>
+    internal sealed class AvroSerializerSettings
     {
         /// <summary>
         ///     Initializes a new instance of the <see cref="AvroSerializerSettings" /> class.
@@ -60,11 +60,6 @@ namespace SolTechnology.Avro.AvroObjectServices.BuildSchema
         internal AvroContractResolver Resolver { get; set; }
 
         /// <summary>
-        /// Gets or sets a serialization surrogate.
-        /// </summary>
-        internal IAvroSurrogate Surrogate { get; set; }
-
-        /// <summary>
         /// Gets or sets a value indicating whether DateTime values will be serialized in the Posix format (as a number
         /// of seconds passed from the start of the Unix epoch) or as a number of ticks.
         /// </summary>
@@ -80,86 +75,5 @@ namespace SolTechnology.Avro.AvroObjectServices.BuildSchema
         ///     The maximum number of items in the schema tree.
         /// </value>
         internal int MaxItemsInSchemaTree { get; set; }
-
-        /// <summary>
-        ///     Gets or sets the known types.
-        /// </summary>
-        internal IEnumerable<Type> KnownTypes { get; set; }
-
-        /// <summary>
-        /// Gets or sets a value indicating whether to use a cache of precompiled serializers.
-        /// </summary>
-        /// <value>
-        ///   <c>True</c> if to use the cache; otherwise, <c>false</c>.
-        /// </value>
-        internal bool UseCache { get; set; }
-
-        /// <summary>
-        ///     Indicates whether the current object is equal to another object of the same type.
-        /// </summary>
-        /// <param name="other">An object to compare with this object.</param>
-        /// <returns>
-        ///     True if the current object is equal to the <paramref name="other" /> parameter; otherwise, false.
-        /// </returns>
-        public bool Equals(AvroSerializerSettings other)
-        {
-            if (other == null)
-            {
-                return false;
-            }
-
-            return this.GenerateSerializer == other.GenerateSerializer
-                && this.GenerateDeserializer == other.GenerateDeserializer
-                && this.UsePosixTime == other.UsePosixTime
-                && this.MaxItemsInSchemaTree == other.MaxItemsInSchemaTree
-                && this.Surrogate == other.Surrogate
-                && this.UseCache == other.UseCache
-                && this.Resolver.Equals(other.Resolver)
-                && this.KnownTypes.SequenceEqual(other.KnownTypes);
-        }
-
-        /// <summary>
-        ///     Determines whether the specified <see cref="System.Object" /> is equal to this instance.
-        /// </summary>
-        /// <param name="obj">
-        ///     The <see cref="System.Object" /> to compare with this instance.
-        /// </param>
-        /// <returns>
-        ///     <c>true</c> if the specified <see cref="System.Object" /> is equal to this instance; otherwise, <c>false</c>.
-        /// </returns>
-        public override bool Equals(object obj)
-        {
-            return this.Equals(obj as AvroSerializerSettings);
-        }
-
-        /// <summary>
-        ///     Returns a hash code for this instance.
-        /// </summary>
-        /// <returns>
-        ///     A hash code for this instance, suitable for use in hashing algorithms and data structures like a hash table.
-        /// </returns>
-        public override int GetHashCode()
-        {
-            unchecked
-            {
-                var hashcode = 83;
-                hashcode = (hashcode * 89) + this.GenerateSerializer.GetHashCode();
-                hashcode = (hashcode * 89) + this.GenerateDeserializer.GetHashCode();
-                hashcode = (hashcode * 89) + this.UsePosixTime.GetHashCode();
-                hashcode = (hashcode * 89) + this.MaxItemsInSchemaTree.GetHashCode();
-                hashcode = (hashcode * 89) + (this.Resolver != null ? this.Resolver.GetHashCode() : 0);
-                hashcode = (hashcode * 89) + (this.Surrogate != null ? this.Surrogate.GetHashCode() : 0);
-                hashcode = (hashcode * 89) + this.UseCache.GetHashCode();
-                if (this.KnownTypes != null)
-                {
-                    hashcode = (hashcode * 89) + this.KnownTypes.Count();
-                    foreach (var knownType in this.KnownTypes)
-                    {
-                        hashcode = (hashcode * 89) + (knownType != null ? knownType.GetHashCode() : 0);
-                    }
-                }
-                return hashcode;
-            }
-        }
     }
 }