+++ src/AvroConvert/AvroObjectServices/Write/Resolver.cs
@@ -15,9 +15,8 @@
 */
 #endregion
 
-using System;
-using SolTechnology.Avro.AvroObjectServices.Schema;
-using SolTechnology.Avro.AvroObjectServices.Schema.Abstract;
+using SolTechnology.Avro.AvroObjectServices.Schemas;
+using SolTechnology.Avro.AvroObjectServices.Schemas.Abstract;
 using SolTechnology.Avro.AvroObjectServices.Write.Resolvers;
 using SolTechnology.Avro.Features.Serialize;
 using SolTechnology.Avro.Infrastructure.Exceptions;
@@ -45,6 +44,7 @@ namespace SolTechnology.Avro.AvroObjectServices.Write
         private static readonly Decimal Decimal;
         private static readonly Duration Duration;
         private static readonly TimestampMilliseconds TimestampMilliseconds;
+        private static readonly TimestampMicroseconds TimestampMicroseconds;
 
         static Resolver()
         {
@@ -61,6 +61,7 @@ namespace SolTechnology.Avro.AvroObjectServices.Write
             Decimal = new Decimal();
             Duration = new Duration();
             TimestampMilliseconds = new TimestampMilliseconds();
+            TimestampMicroseconds = new TimestampMicroseconds();
         }
 
         internal static Encoder.WriteItem ResolveWriter(TypeSchema schema)
@@ -95,6 +96,8 @@ namespace SolTechnology.Avro.AvroObjectServices.Write
                                 return Decimal.Resolve((DecimalSchema)logicalTypeSchema);
                             case LogicalTypeSchema.LogicalTypeEnum.TimestampMilliseconds:
                                 return TimestampMilliseconds.Resolve((TimestampMillisecondsSchema)logicalTypeSchema);
+                            case LogicalTypeSchema.LogicalTypeEnum.TimestampMicroseconds:
+                                return TimestampMicroseconds.Resolve((TimestampMicrosecondsSchema)logicalTypeSchema);
                             case LogicalTypeSchema.LogicalTypeEnum.Duration:
                                 return Duration.Resolve((DurationSchema)logicalTypeSchema);
                         }