+++ docs/CHANGELOG.md
@@ -291,7 +291,12 @@ START OF THE PROJECT
 **v. 3.2.8 (01.06.22)**
 - Added support for Fixed and Map in GenerateModel
 
+**v. 3.2.9 (22.06.22)**
+- Added support for union of Records
+- Fixed Avro2Json timestamp serialization and deserialization
+- Added Generic Json2Avro feature as foundation for dynamic one
+
 \
 **v. 3.3.0 (xxx)**
 - Added support for short (int16) C# type
-- Allowed conversion for different read and write number C# types (ex: int -> long)
+- Allowed conversion for non-matching read and write number C# types (ex: int -> long)